norm OK, path 지우고 ls 실행됨, 추가 테스트 필요
