+++ minishell_parsing_env.c
@@ -6,66 +6,13 @@
 /*   By: jimchoi <jimchoi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/10 20:42:03 by jimchoi           #+#    #+#             */
-/*   Updated: 2024/06/15 15:06:38 by jimchoi          ###   ########.fr       */
+/*   Updated: 2024/06/15 16:13:47 by jimchoi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 #include "minishell_exec.h"
 #include "minishell_parsing.h"
-
-int	env_split_count(char *s)
-{
-	int	i;
-	int	count;
-
-	i = 0;
-	count = 0;
-	while (s[i])
-	{
-		if (s[i] != '$' && s[i] != '\0')
-		{
-			while (s[i] != '$' && s[i] != '\0')
-				i++;
-		}
-		else
-		{
-			i++;
-			while (ft_isalnum(s[i]) != 0 || s[i] == '?')
-				i ++;
-		}
-		count++;
-	}
-	return (count);
-}
-
-char	**make_env_result(char **result, char *s, int i, int idx)
-{
-	int		len;
-
-	while (s[i])
-	{
-		len = 0;
-		if (s[i] != '$' && s[i] != '\0')
-		{
-			while (s[i + len] != '$' && s[i + len] != '\0')
-				len++;
-			result[idx] = ft_substr(s, i, len);
-			i += len;
-		}
-		else
-		{
-			len ++;
-			while (ft_isalnum(s[i + len]) != 0 || s[i + len] == '?')
-				len ++;
-			result[idx] = ft_substr(s, i, len);
-			i += len;
-		}
-		idx ++;
-	}
-	result[idx] = 0;
-	return (result);
-}
 
 char	**env_split(char *s)
 {
@@ -92,24 +39,24 @@ void	env_var_transform(char **result, t_process *prcs)
 	char	*temp;
 
 	i = 0;
+	temp = ft_strdup(result[i]);
+	free(result[i]);
 	node = ft_envpfind(prcs->envp, (result[i] + 1));
-	if (ft_strncmp(result[i], "$?", 3) == 0)
+	if (ft_strncmp(temp, "$?", 3) == 0)
 		result[i] = ft_itoa(prcs->envp->status);
 	else if (node != NULL)
-	{
-		temp = result[i];
 		result[i] = ft_strdup(node->value);
-		free(temp);
-		temp = NULL;
-	}
-	else if (ft_strncmp(result[i], "$PWD", 5) == 0)
+	else if (ft_strncmp(temp, "$PWD", 5) == 0)
 		result[i] = ft_strdup(prcs->senvp.pwd);
-	else if (ft_strncmp(result[i], "$OLDPWD", 7) == 0)
+	else if (ft_strncmp(temp, "$OLDPWD", 7) == 0)
 		result[i] = ft_strdup(prcs->senvp.oldpwd);
-	else if (ft_strncmp(result[i], "$HOME", 6) == 0)
+	else if (ft_strncmp(temp, "$HOME", 6) == 0)
 		result[i] = ft_strdup(prcs->senvp.home);
+	else if (ft_strncmp(temp, "$", 2) == 0)
+		result[i] = ft_strdup("$");
 	else
 		*result = ft_strdup("");
+	free(temp);
 }
 
 void	expand_env_string(char **line, t_process *prcs)
